+++ apps/tailwind-components/app/utils/fetchSetting.ts
