+++ apps/tailwind-components/app/utils/fetchSettings.ts
